refactor: rename display classes
