+++ examples/bare_minimum/bare_minimum.ino
@@ -5,11 +5,11 @@
 
 //Pick your panel:	(uncomment one)
 
-	//Heltec_154_V2	display(/* DC PIN */  8, /* CS PIN */  10, /* BUSY PIN */ 7);
+	//DEPG0150BNS810 display(/* DC PIN */  8, /* CS PIN */  10, /* BUSY PIN */ 7);
 
-	//Heltec_213Red_V2 display(/* DC PIN */  8, /* CS PIN */  10, /* BUSY PIN */ 7);
+	//QYEG0213RWS800 display(/* DC PIN */  8, /* CS PIN */  10, /* BUSY PIN */ 7);
 
-	//Heltec_290_V2 display(/* DC PIN */  8, /* CS PIN */  10, /* BUSY PIN */ 7);
+	//DEPG0290BNS800 display(/* DC PIN */  8, /* CS PIN */  10, /* BUSY PIN */ 7);
 
 void setup() {
 	//Get everything ready
