+++ examples/windowed/windowed.ino
@@ -1,9 +1,9 @@
 //---------------------------------------------------------------------
 //Which display are you using?  (uncomment one)
 
-//#define   USING_Heltec_154_V2
-//#define   USING_Heltec_213Red_V2
-//#define   USING_Heltec_290_V2
+//#define   USING_DEPG0150BNS810
+//#define   USING_QYEG0213RWS800
+//#define   USING_DEPG0290BNS800
 
 
 //Where is your display connected?
@@ -14,12 +14,12 @@
 
 
 //(Example automatically picks the correct class and sample image)
-#if     defined USING_Heltec_154_V2
-            #define     PANEL_CLASS     Heltec_154_V2    
-#elif   defined USING_Heltec_213Red_V2
-            #define     PANEL_CLASS     Heltec_213Red_V2    
-#elif   defined USING_Heltec_290_V2
-            #define     PANEL_CLASS     Heltec_290_V2    
+#if     defined USING_DEPG0150BNS810
+            #define     PANEL_CLASS     DEPG0150BNS810    
+#elif   defined USING_QYEG0213RWS800
+            #define     PANEL_CLASS     QYEG0213RWS800    
+#elif   defined USING_DEPG0290BNS800
+            #define     PANEL_CLASS     DEPG0290BNS800    
 #endif
 //----------------------------------------------------------------------
